Anonymous services can now be serialized.
If the following conditions are meet:
 1- Bigloo version >= 2.8b
 2- The serialization of the service is lesser than (hop-max-url-length)
Then, this service is no longer registered in a global uncollected table.
Instead, it is serialized and sent to the client.
On return, i.e., when the service is invoked, md5sum hashing and HOP-SESSION
ensure the corretness of the service and its validity with respect to the
current session.
In order to disable this feature, one might configure:
  (hop-max-url-length-set! 0)
This feature is highly experimental and should be careful tested.