Fix Scheme->JS compilation, fix host aliasing.
The function COMPILE-HOP-CLIENT was not returning the compiled expression!
This was breaking the REPL demo of the home page.

The HOP host aliasing was checking after the IP number comparison. This was
an error because this was forcing the computation of (host "hop") which,
depending on the local DNS, was raising an error.