Add support for XHTML.
This has implied huge changes. In particular, the HTTP-RESPONSE-HOP instances
now hold an additional value representing the back-end (e.g., html or xhtml)
they are targeting. The function XML-WRITE takes this back-end as an extra
parameter. This technique dissociates the targets from the answers. A benefit
is that a same XML tree can be dumped a first in HTML and a second time
in XHTML.

The class XML-BACKEND is implements the back-end. It is not supposed to
be instantiate by users even thought this is possible.