Start fixing a subtle bug on HSS loading.
The HSS file are evaluated in the current module of the thread
that loads them. This is obviously wrong. I have change this in order
to load them in a fresh module that does not export anything. If a HSS
wants to export something, it has to declare its own module. I think that
this is the first step of a more general fix about HSS loading. I think that
a problem remains due to parallel (multi-threading) load of HSS files.