Hooks are now closed as filters when entering the Hop loop.
This has imposed to re-write the WEBFILTER weblet. This weblet is no longer
an autoloaded one. We have added a new field in the weblet.info specification:
the RC field that contains an expression to be evaluated when this file
is parsed. In addition, when the AUTOLOAD value is #f, the weblet is never
autoloaded (this, of course, was motivated by WEBFILTER).