Change in service name lookup.
When a service is not found in the service registry, the new behavior
consists is recursively search all the dirnames of the path the service.
That is, if the service /hop/foo/bar/gee is not found, the service
/hop/foo/bar then /hop/foo are sequentially checked. The first defined
one is executed. The remaining part of the url can be used by the service
itself as arguments.