Roll back the service URL detection.
That is the dirname chain of a URL is no longer scanned in order to
find the appropriate service. That is, for a valid service
  (define-service (foo/bar/gee ....) ...)

The url service selection for hop/foo/bar/gee/hux won't fall back
to /hop/foo/bar/gee. Instead, it will raise an error. This requires
tricking for implementing nice "download" buttons but it is cleaner.

To compensate this, the SERVICE form accepts an extra optional argument:
 :url name