Change the default history recording. Rename password into epassword.
Widget that uses the history (e.g., tree, notepad) now, by default,
saves their history only if the are explicitly asked to or if they
are provided with a user-specified identity.

The weblet password has been renamed into epassword in order to avoid
the collision name with the Bigloo function password (that reads a
password from the console).