The reader translate all its strings into UTF-8.
This is a temporary workaround for the character encoding which is
currently erroneous. In particular, the XML-WRITE never uses its
ENCODING argument while, of course, it should. At this stage it
is unclear to me where the character encoding should be defined
and where characters should be translated. Possible solutions could be
to set the charset in the XML backend, in the Hop response, and
in the XML elements. This will have to fixed, urgently.