Switch from prefix to suffix technique for client side compilation.
Using a dedicated prefix is not possible for client side script because
it breaks the :base attribute of HEAD markup (browsers complement only
the relative files and thus, not those starting with the /hopcc prefix).
Hence, client side compilation is now launch for files ended with a
dedicated suffix (currently ?hop).